# env/swagger/openapi.yaml
openapi: 3.0.2
info:
  title: Swagger React.js sample API
  description: React.jsサンプル用API
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: http://localhost:3000/
tags:
- name: health-check
  description: ヘルスチェック
- name: adminers
  description: 管理者
- name: users
  description: ユーザ
paths:
  /health-check:
    get:
      tags:
      - health-check
      summary: Health Check API
      operationId: health-check
      responses:
        200:
          description: successful
          content:
            application/json:
              schema:
                type: string
                example: OK
  /adminers/login:
    post:
      tags:
      - adminers
      summary: 管理者ログインAPI
      operationId: adminers_login
      requestBody:
        description: ログイン管理者情報
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: test-adminer@example.com
                password:
                  type: string
                  example: password
        required: true
      security: []
      responses:
        201:
          description: successful
          headers: 
            Set-Cookie:
              schema: 
                type: string
                example: auth-cookie-adminer=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 201
                  message: success
        401:
          description: Unauthorized
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 401
                  message: Unauthorized
        500:
          description: Internal Server Error
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 500
                  message: Internal Server Error
  /adminers/logout:
    post:
      tags:
      - adminers
      summary: 管理者ログアウトAPI
      operationId: adminers_logout
      security:
      - cookieAuthAdminer: []
      responses:
        201:
          description: successful
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 201
                  message: success
        401:
          description: Unauthorized
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 401
                  message: Unauthorized
  /adminers/regist:
    post:
      tags:
      - adminers
      summary: 管理者情報登録API
      operationId: adminers_regist
      requestBody:
        description: 管理者情報
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: テスト管理者
                email:
                  type: string
                  example: test-adminer@example.com
                password:
                  type: string
                  example: password
                confirmPassword:
                  type: string
                  example: password
        required: true
      security: []
      responses:
        201:
          description: successful
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 201
                  message: success
        400:
          description: Bad Request
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 400
                  message: [
                    Email already exist,
                    Confirm password not matching
                  ]
                  error: Bad Request
        500:
          description: Internal Server Error
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 500
                  message: Internal Server Error
  /adminers/profile:
    get:
      security:
      - cookieAuthAdminer: []
      tags:
      - adminers
      summary: 管理者情報取得API
      operationId: adminers_profile
      responses:
        200:
          description: successful
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 200
                  message: success
                  data:
                    id: 1
                    name: テスト管理者
                    userId: test-adminer
                    iat: 12345678
                    exp: 12345678
        401:
          description: Unauthorized
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 401
                  message: Unauthorized
  /adminers/refresh-tokens:
    get:
      security:
      - cookieAuthAdminer: []
      tags:
      - adminers
      summary: トークンリフレッシュAPI
      operationId: adminers_refresh-tokens
      responses:
        200:
          description: successful
          headers: 
            Set-Cookie:
              schema: 
                type: string
                example: auth-cookie-adminer=abcde12345; Path=/; HttpOnly
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 200
                  message: success
        401:
          description: Unauthorized
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 401
                  message: Unauthorized
  /users/login:
    post:
      tags:
      - users
      summary: ユーザログインAPI
      operationId: users_login
      requestBody:
        description: ログインユーザ情報
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: test-user@example.com
                password:
                  type: string
                  example: password
        required: true
      security: []
      responses:
        201:
          description: successful
          headers: 
            Set-Cookie:
              schema: 
                type: string
                example: auth-cookie-user=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 201
                  message: success
        401:
          description: Unauthorized
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 401
                  message: Unauthorized
        500:
          description: Internal Server Error
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 500
                  message: Internal Server Error
  /users/logout:
    post:
      tags:
      - users
      summary: ユーザログアウトAPI
      operationId: users_logout
      security:
      - cookieAuthUser: []
      responses:
        201:
          description: successful
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 201
                  message: success
        401:
          description: Unauthorized
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 401
                  message: Unauthorized
  /users:
    post:
      tags:
      - users
      summary: ユーザ情報登録API
      operationId: users_store
      requestBody:
        description: ユーザ情報
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: テストユーザ
                email:
                  type: string
                  example: test-user@example.com
                password:
                  type: string
                  example: password
                confirmPassword:
                  type: string
                  example: password
        required: true
      security: []
      responses:
        201:
          description: successful
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 201
                  message: success
        400:
          description: Bad Request
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 400
                  message: [
                    Email already exist,
                    Confirm password not matching
                  ]
                  error: Bad Request
        500:
          description: Internal Server Error
          content:
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 500
                  message: Internal Server Error
  /users/profile:
    get:
      security:
      - cookieAuthUser: []
      tags:
      - users
      summary: ユーザ情報取得API
      operationId: users_profile
      responses:
        200:
          description: successful
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 200
                  message: success
                  data:
                    id: 1
                    name: テストユーザ
                    userId: test-user
                    iat: 12345678
                    exp: 12345678
        401:
          description: Unauthorized
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 401
                  message: Unauthorized
  /users/refresh-tokens:
    get:
      security:
      - cookieAuthUser: []
      tags:
      - users
      summary: トークンリフレッシュAPI
      operationId: users_refresh-tokens
      responses:
        200:
          description: successful
          headers: 
            Set-Cookie:
              schema: 
                type: string
                example: auth-cookie-user=abcde12345; Path=/; HttpOnly
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 200
                  message: success
        401:
          description: Unauthorized
          content: 
            application/json:
              schema: 
                type: object
                example:
                  statusCode: 401
                  message: Unauthorized
components:
  schemas:
    adminers:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        adminerId:
          type: string
        email:
          type: string
        password:
          type: string
        refreshToken:
          type: string
        refreshTokenExp:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
    users:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        userId:
          type: string
        email:
          type: string
        password:
          type: string
        refreshToken:
          type: string
        refreshTokenExp:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
  securitySchemes:
    cookieAuthAdminer:
      type: apiKey
      in: cookie
      name: auth-cookie-adminer
    cookieAuthUser:
      type: apiKey
      in: cookie
      name: auth-cookie-user